+++ mayan/apps/mime_types/tests/test_backends.py
@@ -5,6 +5,7 @@ from mayan.apps.documents.tests.literals import TEST_FILE_PDF_PATH
 from mayan.apps.testing.tests.base import BaseTestCase
 
 from ..backends.file_command import MIMETypeBackendFileCommand
+from ..backends.perl_file_mimeinfo import MIMETypeBackendPerlFileMIMEInfo
 
 
 @tag('mime_types')
@@ -26,6 +27,37 @@ class MIMETypeBackendFileCommandTestCase(BaseTestCase):
 
     def test_pdf_mime_type_and_encoding(self):
         backend = MIMETypeBackendFileCommand()
+
+        with open(TEST_FILE_PDF_PATH, mode='rb') as file_object:
+            mime_type, encoding = backend.get_mime_type(
+                file_object=file_object, mime_type_only=False
+            )
+
+        self.assertEqual(mime_type, 'application/pdf')
+        self.assertIn(
+            encoding, ('binary', 'unknown-8bit')
+        )
+
+
+@tag('mime_types')
+class MIMETypeBackendPerlFileMIMEInfoTestCase(BaseTestCase):
+    def test_bad_path(self):
+        with self.assertRaises(expected_exception=DependenciesException):
+            MIMETypeBackendPerlFileMIMEInfo(mimetype_path='/tmp/badpath')
+
+    def test_pdf_mime_type_only(self):
+        backend = MIMETypeBackendPerlFileMIMEInfo()
+
+        with open(TEST_FILE_PDF_PATH, mode='rb') as file_object:
+            mime_type, encoding = backend.get_mime_type(
+                file_object=file_object, mime_type_only=True
+            )
+
+        self.assertEqual(mime_type, 'application/pdf')
+        self.assertEqual(encoding, 'binary')
+
+    def test_pdf_mime_type_and_encoding(self):
+        backend = MIMETypeBackendPerlFileMIMEInfo()
 
         with open(TEST_FILE_PDF_PATH, mode='rb') as file_object:
             mime_type, encoding = backend.get_mime_type(